requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.3.4
img2pdf==0.5.1
Pillow==10.2.0
httpx==0.26.0
